A  controller/signupPost.php M  controller/usersGet.php M  view/inc/navOnline.php M  view/userGet.php M  view/usersGet.php

+++ controller/usersGet.php
@@ -1,14 +1,20 @@
 <?php
 if(isset($user['type']) && $user['type']=='admin'){
     if(isAjax()){
-        $where=[
-            'id'=>segment(2)
-        ];
-        $db->delete("users",$where);
-        print json(false);
+        $acao=segment(3);
+        if($acao=='delete'){
+            $where=[
+                'id'=>segment(2)
+            ];
+            $db->delete("users",$where);
+            print json(true);
+        }
     }else{
         $where=[
-            'id[>]'=>0
+            'id[>]'=>0,
+            'ORDER'=>[
+                'name'=>'ASC'
+            ]
         ];
         $data['users']=$db->select("users","*",$where);
         view('usersGet',$data);
